Move raw-representable support to own framework
In our application we have a number of `NS_ENUM` and `NS_OPTION` types
that we would like to embed in `DictionaryConvertible` types. The
built-in support to serialize these in json as their raw representable
value (usually an integer) is in fact never what we want. But since
Cast's prinicipal serialization methods are overloaded to use raw value
converters, our types can not be extended to `DictionaryConvertible` for
a custom representation.

In Scala, we would use objects or packages to control the scope of
extensions such as the raw-representable support so that applications
can choose the functionality they want to enable. I've imitated that
here with frameworks. An application can `import CastRaw` to enable
automatic serialization of raw values, otherwise those methods are
unavailable.

Although this is working as desired, I should flag a few issues:

1. Even when I link only to the `Cast.framework` I'm able to import
`CastRaw`. I don't understand why this is.

2. I haven't yet added raw frameworks for OSX and TV.

3. If it's preferred to import the raw extentions in the existing `Cast`
framework/scope, I could rearrange the frameworks to do that.

Separate compilation units seem like overkill to control scope, I just
don't see an alternative. Ideally the application could always link to
one framework and choose functionality through import statements alone
-- which is exactly what's happening in #1 above, I just can't see why.